Rename from example_project to template_project
