+++ MANIFEST.in
@@ -3,7 +3,6 @@ include LICENSE
 
 exclude *.yaml
 exclude *.yml
-exclude tox.ini
 
 prune docs
 prune tests
