--- README-upper.md
@@ -1,36 +1,0 @@
-# Templates for Reproducible Research Projects in Economics
-
-![MIT license](https://img.shields.io/github/license/OpenSourceEconomics/econ-project-templates)
-[![image](https://zenodo.org/badge/14557543.svg)](https://zenodo.org/badge/latestdoi/14557543)
-[![Documentation Status](https://readthedocs.org/projects/econ-project-templates/badge/?version=stable)](https://econ-project-templates.readthedocs.io/en/stable/)
-[![image](https://github.com/OpenSourceEconomics/econ-project-templates/actions/workflows/main.yml/badge.svg)](https://github.com/OpenSourceEconomics/econ-project-templates/actions/workflows/main.yml)
-[![image](https://codecov.io/gh/OpenSourceEconomics/econ-project-templates/branch/main/graph/badge.svg)](https://codecov.io/gh/OpenSourceEconomics/econ-project-templates)
-[![pre-commit.ci status](https://results.pre-commit.ci/badge/github/OpenSourceEconomics/econ-project-templates/main.svg)](https://results.pre-commit.ci/latest/github/OpenSourceEconomics/econ-project-templates/main)
-[![image](https://img.shields.io/badge/code%20style-black-000000.svg)](https://github.com/psf/black)
-
-This project aims to provide project templates for economists that make it easy to
-produce reproducible research using one or more of the most frequently used programming
-languages in economics (i.e Python, R, Julia, Stata).
-
-Users and curious visitors please take a look at the
-[documentation](https://econ-project-templates.readthedocs.io/en/stable/). This
-repository is for developing the templates rather than using them.
-
-## Contributing
-
-We welcome suggestions on anything: improving the documentation, bug reports, feature
-requests. Please open an
-[issue](https://github.com/OpenSourceEconomics/econ-project-templates/issues) in these
-cases.
-
-If you want to work on a specific feature, we are more than happy to get you started!
-Please [get in touch briefly](https://www.wiwi.uni-bonn.de/gaudecker/personal_cv.html),
-this is a small team so there is no need for a detailed formal process.
-
-## Contributors
-
-@hmgaudecker @timmens @tobiasraabe @mj023
-
-## Former Contributors
-
-@janosg @PKEuS @philippmuller @julienschat @raholler
