+++ paper/template_project_pres.tex
