+++ pyproject.toml
@@ -2,16 +2,19 @@
 requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.0"]
 
 [tool.setuptools_scm]
-write_to = "src/example_project/_version.py"
+write_to = "src/template_project/_version.py"
 
 [tool.pytask.ini_options]
-paths = ["./src/example_project", "./paper"]
+paths = ["./src/template_project", "./paper"]
 pdbcls = "pdbp:Pdb"
 
 [tool.ruff]
-target-version = "py311"
-select = ["ALL"]
+target-version = "py312"
 fix = true
+
+
+[tool.ruff.lint]
+select = ["ALL"]
 extend-ignore = [
     "S101",  # Use of `assert` detected.
     "ANN",  # Missing type annotations
@@ -19,12 +22,12 @@ extend-ignore = [
     "D104",  # Missing docstring in public package (__init__.py files)
 ]
 
-[tool.ruff.per-file-ignores]
+[tool.ruff.lint.per-file-ignores]
 "tests/*" = ["D", "PD011"]
 "task_*.py" = ["ANN", "ARG001"]
 "docs/**/task_*.py" = ["D103", "S603"]
 
-[tool.ruff.pydocstyle]
+[tool.ruff.lint.pydocstyle]
 convention = "google"
