+++ setup.cfg
@@ -1,7 +1,7 @@
 [metadata]
-name = example_project
+name = template_project
 description = Example project
-long_description = file: README-upper.md
+long_description = file: README.md
 long_description_content_type = text/markdown
 url = https://github.com/OpenSourceEconomics/econ-project-templates
 author = Hans-Martin von Gaudecker
@@ -38,4 +38,4 @@ where = src
 
 [check-manifest]
 ignore =
-    src/example_project/_version.py
+    src/template_project/_version.py
