+++ src/template_project/analysis/model.py
