+++ tests/analysis/test_predict.py
@@ -2,8 +2,8 @@
 
 import pandas as pd
 import pytest
-from example_project.analysis.predict import predict_prob_by_age
 from pandas.testing import assert_frame_equal
+from template_project.analysis.predict import predict_prob_by_age
 
 
 @pytest.fixture()
