+++ tests/data_management/test_clean_data.py
@@ -1,9 +1,9 @@
 import numpy as np
 import pandas as pd
 import pytest
-from example_project.config import TEST_DIR
-from example_project.data_management.clean_data import clean_data
-from example_project.utilities import read_yaml
+from template_project.config import TEST_DIR
+from template_project.data_management.clean_data import clean_data
+from template_project.utilities import read_yaml
 
 
 @pytest.fixture()
